components/recommendation/requirements.txt
fastapi
uvicorn
sentence-transformers
scikit-learn
mysql-connector-python